changed values of the factorials

+++ hw6/src/yj2159_hw6_q3.cpp
@@ -25,7 +25,7 @@ double eApprox(int n) {
         return firstNum+secondNum;
     } else {
         double divNum(0);
-        int fact(1);
+        double fact(1);
         for (int count(1); count<=n; count++){
             fact *= count;
             divNum = 1.0/fact;
